+++ src/OpenSearch/Namespaces/MachineLearningNamespace.php
@@ -1,0 +1,371 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright OpenSearch Contributors
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  The OpenSearch Contributors require contributions made to
+ *  this file be licensed under the Apache-2.0 license or a
+ *  compatible open source license.
+ */
+
+namespace OpenSearch\Namespaces;
+
+use OpenSearch\Namespaces\AbstractNamespace;
+
+/**
+ * Class MachineLearningNamespace
+ *
+ * This class implements common functionality from the
+ * ML Commons API. Please refer to that documentation
+ * for more details about each API operation.
+ *
+ * @link https://opensearch.org/docs/latest/ml-commons-plugin/api/index/
+ */
+class MachineLearningNamespace extends AbstractNamespace
+{
+    /**
+     * $params['body']             = (string) The body of the request (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function createConnector(array $params = []): array
+    {
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Connectors\CreateConnector');
+        $endpoint->setParams($params);
+        $endpoint->setBody($body);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the connector (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function getConnector(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Connectors\GetConnector');
+        $endpoint->setParams($params);
+        $endpoint->setId($id);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['body']             = (string) The body of the request
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function getConnectors(array $params = []): array
+    {
+        if (!isset($params['body'])) {
+            $params['body'] = [
+              'query' => [
+                'match_all' => new \StdClass(),
+              ],
+              'size' => 1000,
+            ];
+        }
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Connectors\GetConnectors');
+        $endpoint->setBody($body);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the connector (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function deleteConnector(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Connectors\DeleteConnector');
+        $endpoint->setParams($params);
+        $endpoint->setId($id);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['body']             = (string) The body of the request (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function registerModelGroup(array $params = []): array
+    {
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\ModelGroups\RegisterModelGroup');
+        $endpoint->setParams($params);
+        $endpoint->setBody($body);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['body']             = (string) The body of the request
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function getModelGroups(array $params = []): array
+    {
+        if (!isset($params['body'])) {
+            $params['body'] = [
+              'query' => [
+                'match_all' => new \StdClass(),
+              ],
+              'size' => 1000,
+            ];
+        }
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\ModelGroups\GetModelGroups');
+        $endpoint->setBody($body);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the model group (Required)
+     * $params['body']           = (array) The body of the request (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function updateModelGroup(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\ModelGroups\UpdateModelGroup');
+        $endpoint->setParams($params);
+        $endpoint->setBody($body);
+        $endpoint->setId($id);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the model group (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function deleteModelGroup(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\ModelGroups\DeleteModelGroup');
+        $endpoint->setParams($params);
+        $endpoint->setId($id);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['body']             = (string) The body of the request (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function registerModel(array $params = []): array
+    {
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Models\RegisterModel');
+        $endpoint->setParams($params);
+        $endpoint->setBody($body);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the model (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function getModel(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Models\GetModel');
+        $endpoint->setParams($params);
+        $endpoint->setId($id);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['body']             = (array) The body of the request
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function getModels(array $params = []): array
+    {
+        if (!isset($params['body'])) {
+            $params['body'] = [
+              'query' => [
+                'match_all' => new \StdClass(),
+              ],
+              'size' => 1000,
+            ];
+        }
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Models\GetModels');
+        $endpoint->setParams($params);
+        $endpoint->setBody($body);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the model (Required)
+     * $params['body']           = (string) The body of the request
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function deployModel(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Models\DeployModel');
+        $endpoint->setParams($params);
+        $endpoint->setId($id);
+        if ($body) {
+            $endpoint->setBody($body);
+        }
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the model (Required)
+     * $params['body']           = (string) The body of the request
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function undeployModel(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Models\UndeployModel');
+        $endpoint->setParams($params);
+        $endpoint->setId($id);
+        if ($body) {
+            $endpoint->setBody($body);
+        }
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the model (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function deleteModel(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Models\DeleteModel');
+        $endpoint->setParams($params);
+        $endpoint->setId($id);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the model (Required)
+     * $params['body']           = (string) The body of the request
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function predict(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $body = $this->extractArgument($params, 'body');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Models\Predict');
+        $endpoint->setParams($params);
+        $endpoint->setId($id);
+        $endpoint->setBody($body);
+
+        return $this->performRequest($endpoint);
+    }
+
+    /**
+     * $params['id']             = (string) The id of the task (Required)
+     *
+     * @param array $params Associative array of parameters
+     *
+     * @return array
+     *   The response.
+     */
+    public function getTask(array $params = []): array
+    {
+        $id = $this->extractArgument($params, 'id');
+        $endpointBuilder = $this->endpoints;
+        $endpoint = $endpointBuilder('MachineLearning\Tasks\GetTask');
+        $endpoint->setParams($params);
+        $endpoint->setId($id);
+
+        return $this->performRequest($endpoint);
+    }
+
+}